copy to remote alternative (#5)
Stop using CopyToRemote as it makes assumptions about file creation time not compatible with text generation involving inputs from secrets or resource outputs.

replace with a printf like approach running on the target.

The local temp file (for previous generation) is no longer needed. This refactoring is a prerequisite to move to Jinja templates for code generation.